Fixed a 'leaking sockets' issue wherein the client was creating more sockets than specified by the configuration.
The root cause was that when connection or protocol negotiation requests in ClientRequestExecutorFactory timed out, we weren't closing the socket before throwing the error. Because the sockets are handled asynchronously, the socket would eventually connect and then be abandoned. Nonetheless, the socket was taking up a thread on the server (using the thread-per-connection socket store implementation). On node bannage (and subsequent pool recycling) this lead to a large number of these "zombie" sockets which left "live" sockets with no available threads on the server.