Add Ant tasks to measure the code coverage of the unit tests using the Cobertura library (http://cobertura.sourceforge.net/).  A nice introductory article can be found at http://www.ibm.com/developerworks/java/library/j-cobertura/
The results are reported both in HTML format for humans, and in an XML format
suitable for other tools, such as the Cobertura plugin for the Hudson continuous
integration server.

To view the report, run "ant coverage" and then view the
dist/coverage/html/index.html file in your browser.