Avro resolving serialization
New abstract class, AvroResolvingSerializer, provides more efficient Avro
serialization than was possible before. Instead of writing out the Schema along
with every entry in a Store, a 4 byte version is stored at the beginning of
every entry.

The version number is the hashCode of the string representation of the Schema,
i.e., Schema.parse(avroSchemaAsString).toString().hashCode(). The conversion of
Schema back to String is to capture any modification the Avro Schema parser
does to the schema.

Schemas are managed in stores.xml just like the JSON schemas. Internally, the
AvroResolvingSerializer uses a Map of the hashCode->Schema rather than
Integer->String like the SerializationDefinition provides. Two new serializer
types have been added to the DefaultSerializerFactory:

 * avro-specific-resolving
 * avro-generic-resolving

It it up to the user to maintain their schemas in stores.xml