Minor code clean up based on review feedback
Protected all logger.(debug|info) statements I added with an is(Debug|Info)Enabled() check.

Made AsyncRecoveryFailureDetector less verbose. When it polls a server to see if it is available, it now prints out a clean INFO level message. It had been printing out a WARN level message *with* a stacktrace that made this expected behavior look much scarier than it really is.