Removed maxTries option from Rebalancing
maxTries was a hack to work around problems with the scheduling of
rebalancing tasks. The new scheduler (is supposed to) ensure that no
node is doing more than one rebalancing thing (stealing or donating) at
time. Therefore, not being able to schedule an asynchronous rebalancing
task (stealer-based) on a node is a big problem and should not be
retried.