Tweaks to Rebalance Controller and Scheduler
Changed '==' to '>=' comparison to control parallelism permitted by
the scheduler. This is defensive programming. Under failure, the '=='
was insufficient. We need to debug/fix that separately. For now, this
change adds some more defense.

Changed some HashSet's to TreeSet's so that they print out in
order. (No perf impact since sets are all small.)

Additional print outs from scheduler so that current set of node (ids)
with work is in the log from the controller/scheduler.

Include donor node in logger.info from StealerBasedRebalanceTask
constructor. This output is now sufficient to translate a task id to a
specific stealer/donor pair in a single line (and all the
partition-stores in follow on lines).