Fix the new admin metadata check command
This is the redo of the fix
https://github.com/voldemort/voldemort/pull/177
Copied description

1) If check metadata fails it says true or false but not the
actual failure.
2) Check metadata sometimes incorrectly reports failure depending
on the order in which stores are returned.
3) Check metadata will reprot success if a store is completely missing
from one of the nodes. It does not track number of times retrieved.

With this commit the check metadata is done using a map.

For stores check
The map will be from storeName to (map from store defintion
to nodeName)

for other
The map will be from object of that key to nodename.

This map will be used for consistency check. This map will group non
matching information to groups for easier analysis. The lenght of the
value in the map will be evaluated for missing information, solving the
problem 2 and 3 listed above