Add overwrite option to fork lift tool
If the key already exists in the destination cluster, forklift tool
will not overwrite the key. With the intent of supporting read only
cluster data to be fork lifted to the read/write cluster, overwrite
option is being added. If the newly forklifted data does not contain
the entire key set from the old, then the new data will be the mix
of old and new. The old data should expire sooner than the new one
enforced by the retention policy.