use maven central for gradle dependencies
1) The jars are directly consumed from libs directory which makes
upgrading the dependencies harder. Also in some cases the transitive
dependencies are not copied correctly which causes issues.
2) Seperated the libs to private-lib and public-lib. private-lib will
contain jars for which I could not find the same version in the maven
central repository. The public-lib are available in maven central
and are not dependent upon by gradle. But still kept for ant
compatibility.
3) As part of the build (ant/gradle) all the dependent libs are
copied into the libs. This is to make the bin/*.sh work. Also there
is some EasyVPL code which depends on the jar in library.
4) Since the classpath is going to be different both the eclipse and
idea will break. But after the download you can run
./gradlew eclipse and ./gradlew idea to regenerate the correct project/
classpath to make it work again.