Exceptions goes uncounted
Get does parallel requests first and if it fails/times move on to the
serial requests. Exceptions are tracked if the parallel requests fail or
timesout after the pipeline is finished. But if the Serial requests are
processing then the exception goes unnoticed. This causes the node to
stay available for a long time affecting the latency and error rates.

This fix reduces the time window an exception goes unnoticed. I thought
about moving the error handling inside the callback and not have it at 2
places. But not sure what will be the perf impact of the error tracking
code, could be negligible but saving that for a later time.