gradle build fails on mac
some of the mac is forced to use Java 8 and Scala plugin does on
Gradle 1.12 does not work with Java 8. Gradle has fixed the issue on
2.0 version.

Added a task to cleanup the config directory files so that changes to
stores.xml will take effect.

Scala build intermittently fails with the error on mac

FAILURE: Build failed with an exception.

    What went wrong:
    Could not resolve all dependencies for configuration ':zinc'.
    > Could not find org.scala-lang:scala-library:2.10.2.

This is caused by incremental zinc compilation. Given the fact that
scala is very rarely used the intermittent failure is very hard to
workaround.

Added a task for cleaning up config folder so that changes in stores.xml
will take effect.