Directly write to socket instead of NIO
Directly write to socket instead of NIO doing the write for you.
This will avoid one round trip for the selector which will improve the
latency and also throughput. pending VPL perf lab runs.
Admin still uses NIO for streaming writes as I do not have time to
verify it and I see less return of value for Admin writes.