ZoneShrinkage endToendTest fails
Zone is shrinked when the nodes receive get/put traffic.
The bootstrap URL is set to node 1 and node 0 is removed at the end
of zone shrinkage. Zone shrinkage metadata is written to nodes in order.
So first node 0 receives the update, that it is no longer part of the
cluster. But before node 1 is updated, the bootstrap code retries to
bootstrap from node 1. This is a race condition between admin client
setting correct metadata on all nodes and the client threads refreshing
the old value from other nodes. The problem is update metadataversion
properties fail as the node 0 is not longer a part of the cluster.

The race condition will be avoided if the bootstrap url and the node
in which we first write the metadata change are one and the same.
This is what this specific fix does, by making sure node 0 is not removed
during zone shrinkage.