slops are not pushed when cluster state changes
Slop pusher caches the cluster state from the server startup.
When the cluster state changes it is not refreshed.
When the node is marked down because of failures, since the failure
detector does not know about these new nodes, it never polls them back
for availability. The node is permanently marked down, until the server
reboots in which case the cluster is refreshed.

Added a way to run integration tests from gradle, but they are still not
part of any junit, junitAll or other tests.