Delete Keys CLI
Add a tool Delete Keys CLI. This tool reads from a keyfile and deletes
the key from the supplied stores. The tools are considered to be in
human readable format and conversion will be attempted to the
appropriate key.

First of all understand that

The tool also supports the following options
  1) --delete-all-versions. If you have more than one value with
conflicting versions, the tool will fail, because it may not have the
value schema to de-serialize the value and resolve the conflict. The
conflict resolution needs to happen before the key is deleted.
  2) --nodeid <> --admin-url <>. If you want to delete keys only from a
particular node. Use the above options. It is useful when you delete the
keys and if a node went down, you want to rerun the tool with that
option.
  3) --find-keys-exist <> . After the delete you can run with this
option to find if any of the keys exist. If the keys are found the tool
dumps the version of each of the keys. The tool waits for the number of
keys from each store before it completes.