Cluster state equals and node names
Cluster.equals returns true when the outline of the cluster is equal
and when the node names are different. But this is not what we want in
most cases. Let us say that you have a non-zoned cluster, on 2 different
data centers. They are going to have the same node ids, same partitions
but they are not equal.

Even within the same data center if one cluster.xml has node 1 as A and
other cluster.xml has node 1 as B, they are different but currently
returns true. So modified the node equivalence to verify the equality.