Ignore the config data directory in tar
When you run the local test server and if you try tar ,
unknowingly you are packing all the data files which takes a long
amount of time. So ignoring the data directory in the config
folder.

Sometimes eclipse messes up and drops .gradle and others in the bin
directory, so ignoring them in the archive as well.