Fix Log message
HdfsFile does not have toString method which causes object id
to be printed in the log message, it broke the script we had
for collecting the download speed. Although speed can be calculated
better now using the stats file, but that is a separate project.

Added number of directories being downloaded, files in addition to size.
This will help to track some more details, as the files if not exist,
dummy files are created in place.

Renamed HDFSFetcherAdvancedTest to HdfsFetcherAdvancedTest to keep it in
sync with other naming conventions.