Add heart beat for NIO Selector
On Voldemort Server, when there is a disk crash the selector
gets hung. Acceptor still keeps on accepting sockets and
after a while the file descriptor limit is reached
and everything is in a hung state. No JVM stack or heap dump.

Now a selector heart beat is added and if the selector is past the
max heart beat time ( defaults to 3 minutes) the acceptor will
stop assigning connections to this selector. If all selectors
are unhealthy ( past max heart beat time) acceptor will close
the socket so that client can recover from these errors faster.