Fix the white space changes
The previous refactor was done from my mac book which did not
replace the tabs with spaces. This messed up lot of the editing.
Instead of re-doing the change with spaces, I just formatted the code
which is easier and no re-verification is required.

you can review teh commit by adding ?w=1 on github url
or use the git diff -w if you are using the command line to ignore
the whitespaces and there are not many changes.