Update metadata version of cluster.xml
At times metadata version on a cluster.xml and stores.xml drifts
The next metadata update instead of consolidating these versions
updates at few places and ignores at few other places.

This causes the client to not re-bootstrap correctly when the
cluster.xml or stores.xml is changed.

Now when a cluster.xml is changed, the version is synchronized
across the cluster to let the clients auto rebootstrap.