Client may not rebootstrap on stores update
Stores metadata version can drift from one server to other
and store create/modify/delete may not cause the client
to rebootstrap.

This fix merges the VectorClock on all the nodes to be updated
so that the Stores version will be updated correctly.

The old methods which does not take parameters are removed
and the new method is the only public method now.