Support for expanding store and view names
General idea is to make the stores.xml less boiler plate by allowing
identical store definitions to to be expressed as a single `<store>`
element with `<name>`` tag in the form `prefix-{name1,name2}`. This will
result in two stores being defined `prefix-name1` and `prefix-name2`
each with the same store definition.

Views get a similar treatment but with a regex matching against store
names and expanding the name field with the match so

```
    <view>
        <name>prefix-$1-view</name>
        <view-of>prefix-(.*)</view-of>
        <view-of-is-regexp>true</view-of-is-regexp>
```

Will match the above stores `prefix-name1` and `prefix-name2` and
generate views `prefix-name1-view` and `prefix-name2-view` backed by the
corresponding store.