BouncyCastle jar is required even when not enabled
BouncyCastle jar is referenced from VoldemortServer and when
VoldemortServer class is loaded, all its references are resolved
which causes it to fail with ClassLoader error.

Added one more level of indirection to avoid the direct reference
and hence BouncyCastle is not required to be available in the class
path when not enabled.