Add File List and Length based check for RO
1) Currently Stores are validated only for the definition and by
doing a random get against any of the partition.

Now for RO stores, the files are validated by their replication
factor.

If RF=1, a line will appear that this store consistency can't be
validated.

Store xyz has replication factor of 1, skipping
consistency check across nodes

If RF > 1, a file should have exactly the number of copies as RF.
If the file lengths does not match, a warning is printed.

 Store abc File ReadOnlyFile [name=97_0_0.data,
size=0] is expected in 2 nodes, but present only in [Node
localhost:6666 [id 0]]
    Store abc File ReadOnlyFile
[name=97_1_0.data, size=408351] is expected in 2 nodes, but present only
in [Node localhost:6669 [id 1]]

If the file is missing the following warning will appear but only once.

The error reporting is very noisy, this is to alert for the presence of
error. So thrown together quickly without considering the ease of use.

Verified backward and forward compatibility of the change.