BnP job corrupts store when compression mismatches
BnP job does not compare compression when making the schema check.
One store was created with compression enabled on the first run
and the next run removed it. The store silently corrupted the data.

Now compression is compared and if compression does not match,
it errors out instead of corrupting the data.