HdfsFetcher's AdminClient is now configurable.
This AdminClient is used to retrieve quota information from within
the HdfsFetcher. It is a bit messy and we probably should not be
using an AdminClient at all for this, but for now this is just a
quick fix to ensure that we can actually configure this server-side
AdminClient with parameters from VoldemortConfig.