Resolved concurrency push conflict and add unit test,  refactor quirky AdmintClient used in HdfsFetcher, and add hyperlink in README.md
1. It was very likely to cause content conflict when multiply jobs try to push to
the same store simultaneously. We add a hashmap in AdminRequestHandler
to check whether a store is currently fetching or not. And if so, we
block the rest of fetching request and throw an exception.

2. Added a unit test case under StoreSwapperTest.

3. The HdfsFetcher used to create an admintClient instance and asked
itself "diskQuotaSizeInKB". This was a quirky used adminClient. We
removed it and passed quota size directly from AdminRequestHandler.

4. Add a hyperlink in README.md to "A quick git guild for people who
want to make contributions to Voldemort".