Use empty stores for Replace Node CLI test
This should have caught a bug, which is already fixed
in the last commit.

There are still couple of bugs in the Offline Mode.

AdminClient storeOps uses socketPort, which will be down in the
offline mode. AdminPort supports full client operations and hence
it should have used the AdminPort.

AdminClient on the voldemort server uses cluster for bootstrapping
which uses the client port again. This is problematic when the node
is node 0. It should use the Admin Port for bootstrapping.

Those bugs are in backlog will fix them later.