Set Metadata version node by node
Previously Metadata version is set using put operation for
entire cluster. As soon as one node succeeds, any failure
is silently ignored. This caused the cluster to fall out of sync
with clients.

Now the Metadata version is set node by node and errors are reported
so the operator knows of any issues.

Fixed the Admin Command line tools to use the New APIs.