Client shell throws error on closing
Admin and the Client factory uses the same suffix, which tries
to delete the already de-registered JMX. Set the different identifier
for admin.

This is just a minor annoyance when you close the shell, it throws
an exception. It is just restricted to voldmeort client shell. The
metrics are not used by any one and overwriting the metrics is a non
issue.