Lower the node unavailable error to debug
When async connect fails, selector reports the error and it is cached in
memory. Next connect call will get the error if it happens within twice
the timeout period.

The logs are logged at the info level, which spams the logs when the
server is unavailable for extended period of times. Now it is dropped
down to debug level.