Don't use Properties(properties) constructor
Properties(properties) constructor has a different behavior than the one
intended.

http://stackoverflow.com/questions/2004833/how-to-merge-two-java-util-properties-objects

>>> copy/pasted text <<<

However, if you treat it like a Map, you need to be very careful with
this:

new Properties(defaultProperties);
This often catches people out, because it looks like a copy constructor,
but it isn't. If you use that constructor, and then call something like
keySet() (inherited from its Hashtable superclass), you'll get an empty
set, because the Map methods of Properties do not take account of the
default Properties object that you passed into the constructor. The
defaults are only recognised if you use the methods defined in
Properties itself, such as getProperty and propertyNames, among others.