Detect and Validate node Ids
Problem:
   Voldemort server takes node id as configuration parameters.
It relies on the node id to identify its role in the cluster.
But most production deployments has only one voldemort server
per host in a cluster. Under these conditions the deduction
of node ids can be automated. In most of the typical production
deployments, only node id changes across the voldemort server
configuration. This causes configuration duplication and difficult
to manage.

Fix :
Node auto detection can be enabled ( disabled by default) by
the property enable.node.id.detection

When enabled host names in cluster.xml will be matched against one of the
FQDN ( InetAddress.getLocalHost().getCanonicalHostName() ).

Validation can also be enabled by the property
validate.node.id . Note that when auto detection is enabled,
validatiion is always run. So enable the validation config, only
if you want to run the validation, but not auto detection.

The Implementation to match is also customizable, but windows and
other operating systems are not considered, but it should work.
It is customizable only for the purpose of writing tests.

Not much work will be required to support those use cases though.

Note:
   Auto detection and validation both will fail when more than one
node is hosted on the same machine. In such cases, both the parameters
should be left in the default disabled state.

Tests will follow in the next commit.