Fix the HintedHandOff flaky tests
1) I made some changes to metadata store in auto detect node id,
and noticed these tests were failing. On investigation the test
failures are caused by using static variables some of which are
modified and based on the ordering they may or may not fail.

2) Removed most of the static usage and made most of them as
parameters.

I still don't completely understand the test as it is quiet complicated,
but sprinkled in some sleep to make sure that slops are registered.

Tests passed successfully on 50 continous runs.