Adhere target/source java compatibility for contrib projects
We are using Java 1.7 on our hadoop cluster, but locally I use JDK 1.8
for building, so contrib got build with 1.8 by default because its
source and target version is not specified explicitly. Consequentially
the resulting jar didn't work in our cluster environment. This
reconfiguration fixes this.

But since the contrib projects don't build anymore with 1.6 source
compatibility I bumped up the configured javac.version to 1.7, even
though two other options are available:

1. Fix contrib to make it 1.6 compatible again
2. Specify different versions for the main and contrib projects

But since Java 1.6 is EOL since quite a while I suggest that no
Voldemort servers with a Java 1.6 runtime should be running anymore and
it should be save to upgrade and keep the configuration simple.