Changed the DeleteAllFailedFetchStrategy so that it affects all nodes.
Previously, the DeleteAllFailedFetchStrategy would only attempt to
delete data from nodes which succeeded in their fetch. In some failure
modes, this is appropriate, but in other cases, it isn't. In any case,
there is no harm in trying to delete data on all nodes, even those
that failed their fetch. This commit makes it so.