BnP now retries fetches when cluster.xml is stale.
Previously, there was a race condition where a BnP job would initialize
its AdminClients at the beginning of the job, and then hang on to that
Cluster state throughout the job. If a maintenance is going on while
the job is running, then it's possible that by the time the job gets to
the "Push" phase, the Cluster representation may be stale. In those
cases, it is possible that a BnP job may attempt to push to a node
which has been swapped out of the cluster. This may cause BnP HA to
trigger even though the cluster is actually healthy at that time.

In order to fix this, two changes are made in this commit:
1. In the VoldemortSwapJob, the AdminClient is constructed from scratch
   rather than being created based on the previous Cluster state. This
   should minimize the window during which it is possible to change
   the cluster.xml and make BnP hit the wrong node, but it does not
   completely eliminate the race condition.
2. In the AdminStoreSwapper, the invokeFetch() code will check if an
   exception is caused by a stale cluster state. If it is, it will
   get a fresh AdminClient and retry the operation. This should totally
   prevent the race condition.

The BnP job will do a limited amount of fetch retries (10 attempts with
30 seconds of wait time between each) and only when hitting soft errors
(i.e.: connection failure, etc.).