Tweaked the AdminClient's currentVersion so that it is not stale.
Previously, there could be a case where an AdminClient is created from
a stale Cluster instance, which would lead to isClusterModified() not
returning the correct result. This was because the AdminClient would
always set its currentVersion to the current time, no matter how long
ago the passed in Cluster instance was originally generated. In cases
where the cluster.xml configuration is altered after the Cluster
instance is constructed, but before the AdminClient is constructed,
then there is a potentially very long window during which the wrong
currentVersion would be set.