BnP now deals gracefully with rolling restarts.
Previously, BnP could not be run during a rolling restart of the servers.
The recommended best practice was to disable fetches during a planned
maintenance. BnP HA does not help either, because it would either cripple
one of the nodes or fail because more than one node failed to fetch.

A previous commit already added a provision so that when BnP job detects
a soft failure, instead of giving up right away, it will retry up to 10
times with 30 seconds of wait time between each attempt.

This commit is the last missing piece to support rolling restarts. If the
BnP job receives an error saying that an async job does not exist or that
the store directory already exists, then it will attempt to delete the
partially downloaded store-version from the server before retrying the
fetch again.