New RECOVERY mode to partially automate node repair following BnP HA.
In the new RECOVERY mode, the server will deny online requests from
clients (just like in OFFLINE mode) but will allow new BnP jobs to
push into the node. After starting the server in this mode, the
operator can kick off a data restoration job to recover the rest of
the missing stores. Meanwhile, the server will periodically check
if any of its symlinks have changed, and if so will load the newly
restored stores as the active version. Finally, when no more stores
are disabled, the server automatically transitions to the ONLINE
mode and resume serving client requests.

The benefits of this new mode are:

1) Less manual work for the operator.
2) No need for a planned BnP outage during the node recovery process.